pyproject.toml
[tool.poetry]
name = "aitestgen"
version = "0.0.1"
description = "Tool to generate simple unit test using openai api"
authors = ["Pablo España <espana.pablo95@gmail.com>"]
readme = "README.md"
packages = [{include = "aitestgen"}]


[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
python-dotenv = "^1.0.0"
openai = "^1.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
aitestgen = "aitestgen.__main__:cli"